feat(image): Add support for `~/assets` alias
