+++ packages/astro/src/assets/vite-plugin-assets.ts
@@ -30,6 +30,18 @@ export default function assets({
 		// Expose the components and different utilities from `astro:assets` and handle serving images from `/_image` in dev
 		{
 			name: 'astro:assets',
+			config() {
+				return {
+					resolve: {
+						alias: [
+							{
+								find: /^~\/assets\/(.+)$/,
+								replacement: fileURLToPath(new URL('./assets/$1', settings.config.srcDir)),
+							},
+						],
+					},
+				};
+			},
 			async resolveId(id) {
 				if (id === VIRTUAL_SERVICE_ID) {
 					return await this.resolve(settings.config.image.service);
