+++ packages/astro/src/core/path.ts
@@ -64,3 +64,7 @@ export function removeQueryString(path: string) {
 	const index = path.lastIndexOf('?');
 	return index > 0 ? path.substring(0, index) : path;
 }
+
+export function isRemotePath(src: string) {
+	return /^(https?:)?\/\//.test(src);
+}
